# specs/001-todo-backend-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Todo Backend API
  description: Secure RESTful API for multi-user Todo operations with JWT-based authentication
  version: 1.0.0
  contact:
    name: API Support
    email: support@todoapp.com

servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://todo-backend-api.com
    description: Production server

paths:
  /api/v1/tasks:
    get:
      summary: List user's tasks
      description: Retrieve a list of tasks belonging to the authenticated user
      tags:
        - Tasks
      parameters:
        - name: completed
          in: query
          description: Filter tasks by completion status
          required: false
          schema:
            type: boolean
        - name: date_from
          in: query
          description: Filter tasks created after this date (ISO format)
          required: false
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: Filter tasks created before this date (ISO format)
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successfully retrieved tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: []

    post:
      summary: Create a new task
      description: Create a new task for the authenticated user
      tags:
        - Tasks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreateRequest'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
      security:
        - bearerAuth: []

  /api/v1/tasks/{task_id}:
    get:
      summary: Get a specific task
      description: Retrieve a specific task by ID for the authenticated user
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to retrieve
          schema:
            type: integer
      responses:
        '200':
          description: Task retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: []

    put:
      summary: Update a task
      description: Update an existing task for the authenticated user (with optimistic locking)
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to update
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdateRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/ConflictError'
      security:
        - bearerAuth: []

    delete:
      summary: Delete a task
      description: Delete a specific task by ID for the authenticated user
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to delete
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: []

  /api/v1/tasks/{task_id}/toggle-complete:
    patch:
      summary: Toggle task completion status
      description: Toggle the completion status of a specific task
      tags:
        - Tasks
      parameters:
        - name: task_id
          in: path
          required: true
          description: ID of the task to toggle
          schema:
            type: integer
      responses:
        '200':
          description: Task completion status toggled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
      security:
        - bearerAuth: []

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    TaskResponse:
      type: object
      required:
        - id
        - title
        - is_completed
        - created_at
        - updated_at
        - user_id
        - version
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Title of the task
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          nullable: true
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          description: Whether the task is completed
          example: false
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2023-10-20T10:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2023-10-20T10:00:00Z"
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
        version:
          type: integer
          description: Version number for optimistic locking
          example: 1

    TaskCreateRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: Title of the task (1-255 characters)
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Optional description of the task
          maxLength: 65535
          example: "Milk, eggs, bread"

    TaskUpdateRequest:
      type: object
      required:
        - version
      properties:
        title:
          type: string
          description: Updated title of the task (1-255 characters)
          minLength: 1
          maxLength: 255
          example: "Buy groceries"
        description:
          type: string
          description: Updated description of the task
          maxLength: 65535
          example: "Milk, eggs, bread"
        is_completed:
          type: boolean
          description: Updated completion status of the task
          example: false
        version:
          type: integer
          description: Current version for optimistic locking
          example: 1

    ProblemDetail:
      type: object
      required:
        - type
        - title
        - status
        - detail
      properties:
        type:
          type: string
          description: URI reference to the type of error
          example: "https://httpstatuses.com/401"
        title:
          type: string
          description: Brief summary of the error
          example: "Unauthorized"
        status:
          type: integer
          description: HTTP status code
          example: 401
        detail:
          type: string
          description: Specific details about this occurrence of the error
          example: "Invalid or missing authentication token"
        instance:
          type: string
          description: URI reference to the specific occurrence
          example: "/api/v1/tasks/123"

  responses:
    BadRequestError:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    UnauthorizedError:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    ForbiddenError:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'

    ConflictError:
      description: Conflict - optimistic lock failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'